+++ packages/client/src/api/github-cloud-replay.api.ts
@@ -1,6 +1,6 @@
 import { maybeEnrichFetchError } from "../errors";
 import { MeticulousClient } from "../types/client.types";
-import { TestRun } from "./test-run.api";
+import type { TestRun } from "@alwaysmeticulous/api";
 
 export interface GetBaseTestRunOptions {
   client: MeticulousClient;
